requirements/base.txt


django==1.9.6

# Configuration
django-environ==0.4.0
django-secure==1.0.1



# Models
django-model-utils==2.5

# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.5.0
redis>=2.10.0



# Your custom requirements go here
django-celery==3.1.17
django-favicon==0.1.2
kombu==3.0.35
Pygments==2.1.3
requests[security]==2.9.1
xmltodict==0.10.1
zeep==0.20.0
celery==3.1.24
django-admin-bootstrapped
djangorestframework
facebook-sdk==1.0.0
Pillow==3.2.0
